+++ fibjs/src/encoding/encoding_iconv.cpp
@@ -166,14 +166,14 @@ result_t encoding_iconv::encode(exlib::string data, exlib::string& retVal)
         retVal = data;
     else if ((m_charset == "ucs2") || (m_charset == "ucs-2")
         || (m_charset == "utf16le") || (m_charset == "utf-16le")) {
-        int32_t n = utf_convert(data.c_str(), (int32_t)data.length(), (exlib::wchar*)NULL, 0);
+        ssize_t n = utf_convert(data.c_str(), data.length(), (exlib::wchar*)NULL, 0);
         retVal.resize(n * sizeof(exlib::wchar));
-        utf_convert(data.c_str(), (int32_t)data.length(), (exlib::wchar*)&retVal[0], n);
+        utf_convert(data.c_str(), data.length(), (exlib::wchar*)&retVal[0], n);
     } else if ((m_charset == "ucs4") || (m_charset == "ucs-4")
         || (m_charset == "utf32le") || (m_charset == "utf-32le")) {
-        int32_t n = utf_convert(data.c_str(), (int32_t)data.length(), (exlib::wchar32*)NULL, 0);
+        ssize_t n = utf_convert(data.c_str(), data.length(), (exlib::wchar32*)NULL, 0);
         retVal.resize(n * sizeof(exlib::wchar32));
-        utf_convert(data.c_str(), (int32_t)data.length(), (exlib::wchar32*)&retVal[0], n);
+        utf_convert(data.c_str(), data.length(), (exlib::wchar32*)&retVal[0], n);
     } else {
         if (m_charset == "binary")
             m_charset = "latin1";
@@ -223,10 +223,10 @@ result_t encoding_iconv::decode(const exlib::string& data, exlib::string& retVal
         retVal = data;
     else if ((m_charset == "ucs2") || (m_charset == "ucs-2")
         || (m_charset == "utf16le") || (m_charset == "utf-16le"))
-        retVal = utf16to8String((const exlib::wchar*)data.c_str(), (int32_t)data.length() / sizeof(exlib::wchar));
+        retVal = utf16to8String((const exlib::wchar*)data.c_str(), data.length() / sizeof(exlib::wchar));
     else if ((m_charset == "ucs4") || (m_charset == "ucs-4")
         || (m_charset == "utf32le") || (m_charset == "utf-32le"))
-        retVal = utf32to8String((const exlib::wchar32*)data.c_str(), (int32_t)data.length() / sizeof(exlib::wchar32));
+        retVal = utf32to8String((const exlib::wchar32*)data.c_str(), data.length() / sizeof(exlib::wchar32));
     else {
         if (m_charset == "binary")
             m_charset = "latin1";
